Remove hasChanged, changedAttributes, previous, previousAttributes. But keep them around for testing.
